server/requirements.txt
databases==0.3.2
fastapi==0.54.1
uvicorn==0.11.5
passlib==1.7.2
python-dotenv==0.13.0
python-jose==3.1.0
pymysql==0.9.3
sqlalchemy==1.3.16